# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: M&S Delivery API
  version: 1.0.0
servers:
  - url: https://api.example.com
paths:
...
  /api/v1/orders:
    get:
      operationId: listOrders
      security: [{bearerAuth: []}]
      responses:
        '200': {description: Role-scoped orders}
    post:
      operationId: createOrder
      security: [{bearerAuth: []}]
      parameters:
        - {in: header, name: Idempotency-Key, required: true, schema: {type: string}}
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/CreateOrderRequest'}
      responses:
        '201':
          description: Order created
          content: {application/json: {schema: {$ref: '#/components/schemas/Order'}}}
  /api/v1/orders/{order_id}:
    get:
      operationId: getOrder
      security: [{bearerAuth: []}]
      parameters:
        - {in: path, name: order_id, required: true, schema: {type: string, format: uuid}}
      responses:
        '200': {description: Order snapshot}
  /api/v1/orders/{order_id}/assign:
    post:
      operationId: assignRider
      security: [{bearerAuth: []}]
      parameters:
        - {in: path, name: order_id, required: true, schema: {type: string, format: uuid}}
      responses:
        '200': {description: Rider assigned}
...
  /api/v1/orders/{order_id}/feedback:
    post:
      operationId: submitFeedback
      security: [{bearerAuth: []}]
      parameters:
        - {in: path, name: order_id, required: true, schema: {type: string, format: uuid}}
      responses:
        '201': {description: Feedback submitted}
...
  /api/v1/uploads/menu-images/authorize:
    post:
      operationId: authorizeMenuImageUpload
      security: [{bearerAuth: []}]
      responses:
        '200': {description: Signed upload URL}
  /api/v1/uploads/menu-images/finalize:
    post:
      operationId: finalizeMenuImageUpload
      security: [{bearerAuth: []}]
      responses:
        '200': {description: Image recorded}
components:
  securitySchemes:
    bearerAuth: {type: http, scheme: bearer, bearerFormat: JWT}
  schemas:
...
    OrderStatus:
      type: string
      enum: [pending, confirmed, assigned, picked_up, on_the_way, delivered, cancelled]
    PaymentStatus:
      type: string
      enum: [unpaid, paid]
...
    CreateOrderItem:
      type: object
      required: [menu_item_id, quantity]
      properties:
        menu_item_id: {type: string, format: uuid}
        quantity: {type: integer, minimum: 1, maximum: 99}
    CreateOrderRequest:
      type: object
      required: [store_id, delivery_address_id, items]
      properties:
        store_id: {type: string, format: uuid}
        delivery_address_id: {type: string, format: uuid}
        notes: {type: string, maxLength: 500}
        items:
          type: array
          minItems: 1
          items: {$ref: '#/components/schemas/CreateOrderItem'}
    Order:
      type: object
      required: [id, status, payment_status, subtotal, delivery_fee, total]
      properties:
        id: {type: string, format: uuid}
        status: {$ref: '#/components/schemas/OrderStatus'}
        payment_status: {$ref: '#/components/schemas/PaymentStatus'}
        subtotal: {type: number}
        delivery_fee: {type: number}
        total: {type: number}
        delivery_id: {type: [string, 'null'], format: uuid}
...